backend/requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
pydantic
python-multipart
pytest